Move management server to a separate directory
