+++ management/impl/root.go
@@ -1,4 +1,4 @@
-package cmd
+package impl
 
 import (
 	"fmt"
@@ -44,13 +44,7 @@ func init() {
 	}
 	rootCmd.PersistentFlags().StringVar(&configPath, "config", defaultConfigPath, "Wiretrustee config file location to write new config to")
 	rootCmd.PersistentFlags().StringVar(&logLevel, "log-level", "info", "")
-	rootCmd.AddCommand(initCmd)
-	rootCmd.AddCommand(addPeerCmd)
-	rootCmd.AddCommand(upCmd)
-	rootCmd.AddCommand(signalCmd)
-	rootCmd.AddCommand(serviceCmd)
-	serviceCmd.AddCommand(runCmd, startCmd, stopCmd, restartCmd) // service control commands are subcommands of service
-	serviceCmd.AddCommand(installCmd, uninstallCmd)              // service installer commands are subcommands of service
+	rootCmd.AddCommand(mgmtCmd)
 }
 
 // SetupCloseHandler handles SIGTERM signal and exits with success
