+++ management/impl/store.go
@@ -1,4 +1,4 @@
-package management
+package impl
 
 // Account represents a unique account of the system
 type Account struct {
